requirements.txt
streamlit
opencv-python
numpy
google-generativeai
python-dotenv
scikit-learn
requests
beautifulsoup4
